requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
pillow==11.0.0
pycryptodome==3.21.0
secretsharing==0.2.6
ttkbootstrap==1.10.1
